Refactor internals of GitHub module
The GitHub module was used as a catch-all for small newtypes that were
used to interact with GitHub's API. As we are planning to create
interactions with GitHub, the need for more as well as more complex
types becomes clear. Instead of adding these to the current module or
mixing styles, the decision has been made to refactor the existing
module and then add anything new inside the github/ directory.
