+++ src/github/mod.rs
@@ -1,0 +1,11 @@
+//! Entities to interact with GitHub
+
+pub use self::app_id::AppId;
+pub use self::github_host::GitHubHost;
+pub use self::private_key::PrivateKey;
+pub use self::webhook_secret::WebhookSecret;
+
+mod app_id;
+mod github_host;
+mod private_key;
+mod webhook_secret;
